feat(tokenizer): change token types to enum
in my mind, this is less efficient. then again, using an IToken meant that every value was a class, or if it was a struct it would be boxed anyways, and whether that has any memory implications has yet to be tested.
regardless, as of right now this is an easy to change thing so if I decide to switch back I can do so, and benchmarking is of more importance at a later date, not now.